refactor: move integration module into integration/ package
Phase 6 of #625 - reorganize project structure:
- battery_controller.py → integration/controller.py
- powerwall_service_client.py → integration/client.py
- Updated all imports and mock patches in tests
